# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HTTPResponse-array_model_Task:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      message:
        type: string
    type: object
  httputil.HTTPResponse-model_Task:
    properties:
      data:
        $ref: '#/definitions/model.Task'
      message:
        type: string
    type: object
  model.Task:
    properties:
      createdAt:
        type: integer
      description:
        type: string
      id:
        type: string
      recurrence:
        $ref: '#/definitions/model.TaskRecurrence'
      scheduledTime:
        type: integer
      status:
        $ref: '#/definitions/model.TaskStatus'
      title:
        type: string
      updatedAt:
        type: integer
      version:
        type: string
    required:
    - recurrence
    - scheduledTime
    - status
    - title
    type: object
  model.TaskRecurrence:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - Once
    - Daily
    - Weekly
    - Monthly
  model.TaskStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - StatusPending
    - StatusInProgress
    - StatusCompleted
    - StatusFailed
host: localhost:8080
info:
  contact: {}
  description: REST API for managing task scheduling
  title: DTS API
  version: 0.1.0
paths:
  /tasks:
    get:
      description: retrieves all tasks
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.HTTPResponse-array_model_Task'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKey: []
      summary: Get all tasks
      tags:
      - tasks
    put:
      description: creates or updates a task
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.HTTPResponse-model_Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKey: []
      summary: Put a task
      tags:
      - tasks
  /tasks/:id:
    delete:
      description: deletes a specific task
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.HTTPResponse-model_Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKey: []
      summary: Delete a task
      tags:
      - tasks
    get:
      description: retrieves a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httputil.HTTPResponse-model_Task'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKey: []
      summary: Get a specific task
      tags:
      - tasks
schemes:
- http
securityDefinitions:
  ApiKey:
    description: User-specific API key
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"